Extract pkg/cloudformation/customresources.
