+++ server/cloudformation/cloudformation.go
@@ -3,14 +3,10 @@
 package cloudformation
 
 import (
-	"bytes"
 	"database/sql"
 	"encoding/json"
 	"fmt"
-	"hash/fnv"
-	"io/ioutil"
 	"net/http"
-	"strconv"
 	"time"
 
 	"golang.org/x/net/context"
@@ -19,7 +15,7 @@ import (
 	"github.com/aws/aws-sdk-go/aws/client"
 	"github.com/aws/aws-sdk-go/service/ecs"
 	"github.com/aws/aws-sdk-go/service/sqs"
-	"github.com/remind101/empire/pkg/base62"
+	"github.com/remind101/empire/pkg/cloudformation/customresources"
 	"github.com/remind101/empire/scheduler/ecs/lb"
 	"github.com/remind101/pkg/logger"
 	"github.com/remind101/pkg/reporter"
@@ -34,198 +30,10 @@ var (
 	ProvisioningGraceTimeout = time.Duration(1 * time.Minute)
 )
 
-// Provisioner is something that can provision custom resources.
-type Provisioner interface {
-	// Provision should do the appropriate provisioning, then return:
-	//
-	// 1. The physical id that was created, if any.
-	// 2. The data to return.
-	Provision(context.Context, Request) (string, interface{}, error)
-
-	// Properties should return an instance of a type that the properties
-	// can be json.Unmarshalled into.
-	Properties() interface{}
-}
-
-type ProvisionerFunc func(context.Context, Request) (string, interface{}, error)
-
-func (fn ProvisionerFunc) Provision(ctx context.Context, r Request) (string, interface{}, error) {
-	return fn(ctx, r)
-}
-
-// withTimeout wraps a Provisioner with a context.WithTimeout.
-func withTimeout(p Provisioner, timeout time.Duration, grace time.Duration) Provisioner {
-	return &timeoutProvisioner{
-		Provisioner: p,
-		timeout:     timeout,
-		grace:       grace,
-	}
-}
-
-type result struct {
-	id   string
-	data interface{}
-	err  error
-}
-
-type timeoutProvisioner struct {
-	Provisioner
-	timeout time.Duration
-	grace   time.Duration
-}
-
-func (p *timeoutProvisioner) Provision(ctx context.Context, r Request) (string, interface{}, error) {
-	ctx, cancel := context.WithTimeout(ctx, p.timeout)
-	defer cancel()
-
-	done := make(chan result)
-	go func() {
-		id, data, err := p.Provisioner.Provision(ctx, r)
-		done <- result{id, data, err}
-	}()
-
-	select {
-	case r := <-done:
-		return r.id, r.data, r.err
-	case <-ctx.Done():
-		// When the context is canceled, give the provisioner
-		// some extra time to cleanup.
-		<-time.After(p.grace)
-		select {
-		case r := <-done:
-			return r.id, r.data, r.err
-		default:
-			return "", nil, ctx.Err()
-		}
-	}
-}
-
-// Possible request types.
-const (
-	Create = "Create"
-	Update = "Update"
-	Delete = "Delete"
-)
-
-// Request represents a Custom Resource request.
-//
-// See http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/crpg-ref-requests.html
-type Request struct {
-	// The request type is set by the AWS CloudFormation stack operation
-	// (create-stack, update-stack, or delete-stack) that was initiated by
-	// the template developer for the stack that contains the custom
-	// resource.
-	//
-	// Must be one of: Create, Update, or Delete.
-	RequestType string `json:"RequestType"`
-
-	// The response URL identifies a pre-signed Amazon S3 bucket that
-	// receives responses from the custom resource provider to AWS
-	// CloudFormation.
-	ResponseURL string `json:"ResponseURL"`
-
-	// The Amazon Resource Name (ARN) that identifies the stack containing
-	// the custom resource.
-	//
-	// Combining the StackId with the RequestId forms a value that can be
-	// used to uniquely identify a request on a particular custom resource.
-	StackId string `json:"StackId"`
-
-	// A unique ID for the request.
-	//
-	// Combining the StackId with the RequestId forms a value that can be
-	// used to uniquely identify a request on a particular custom resource.
-	RequestId string `json:"RequestId"`
-
-	// The template developer-chosen resource type of the custom resource in
-	// the AWS CloudFormation template. Custom resource type names can be up
-	// to 60 characters long and can include alphanumeric and the following
-	// characters: _@-.
-	ResourceType string `json:"ResourceType"`
-
-	// The template developer-chosen name (logical ID) of the custom
-	// resource in the AWS CloudFormation template. This is provided to
-	// facilitate communication between the custom resource provider and the
-	// template developer.
-	LogicalResourceId string `json:"LogicalResourceId"`
-
-	// A required custom resource provider-defined physical ID that is
-	// unique for that provider.
-	//
-	// Always sent with Update and Delete requests; never sent with Create.
-	PhysicalResourceId string `json:"PhysicalResourceId"`
-
-	// This field contains the contents of the Properties object sent by the
-	// template developer. Its contents are defined by the custom resource
-	// provider.
-	ResourceProperties interface{} `json:"ResourceProperties"`
-
-	// Used only for Update requests. Contains the resource properties that
-	// were declared previous to the update request.
-	OldResourceProperties interface{} `json:"OldResourceProperties"`
-}
-
-// Possible response statuses.
-const (
-	StatusSuccess = "SUCCESS"
-	StatusFailed  = "FAILED"
-)
-
-// Response represents the response body we send back to CloudFormation when
-// provisioning is complete.
-//
-// See http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/crpg-ref-responses.html
-type Response struct {
-	// The status value sent by the custom resource provider in response to
-	// an AWS CloudFormation-generated request.
-	//
-	// Must be either SUCCESS or FAILED.
-	Status string `json:"Status"`
-
-	// Describes the reason for a failure response.
-	//
-	// Required if Status is FAILED; optional otherwise.
-	Reason string `json:"Reason"`
-
-	// This value should be an identifier unique to the custom resource
-	// vendor, and can be up to 1Kb in size. The value must be a non-empty
-	// string.
-	PhysicalResourceId string `json:"PhysicalResourceId"`
-
-	// The Amazon Resource Name (ARN) that identifies the stack containing
-	// the custom resource. This response value should be copied verbatim
-	// from the request.
-	StackId string `json:"StackId"`
-
-	// A unique ID for the request. This response value should be copied
-	// verbatim from the request.
-	RequestId string `json:"RequestId"`
-
-	// The template developer-chosen name (logical ID) of the custom
-	// resource in the AWS CloudFormation template. This response value
-	// should be copied verbatim from the request.
-	LogicalResourceId string `json:"LogicalResourceId"`
-
-	// Optional, custom resource provider-defined name-value pairs to send
-	// with the response. The values provided here can be accessed by name
-	// in the template with Fn::GetAtt.
-	Data interface{} `json:"Data"`
-}
-
 // Represents the body of the SQS message, which would have been received from
 // SNS.
 type Message struct {
 	Message string `json:"Message"`
-}
-
-// NewResponseFromRequest initializes a new Response from a Request, filling in
-// the required verbatim fields.
-func NewResponseFromRequest(req Request) Response {
-	return Response{
-		StackId:           req.StackId,
-		RequestId:         req.RequestId,
-		LogicalResourceId: req.LogicalResourceId,
-	}
 }
 
 // sqsClient duck types the sqs.SQS interface.
@@ -248,7 +56,7 @@ type CustomResourceProvisioner struct {
 
 	// Provisioners routes a custom resource to the thing that should do the
 	// provisioning.
-	Provisioners map[string]Provisioner
+	Provisioners map[string]customresources.Provisioner
 
 	client interface {
 		Do(*http.Request) (*http.Response, error)
@@ -260,7 +68,7 @@ type CustomResourceProvisioner struct {
 // sqs client configured from config.
 func NewCustomResourceProvisioner(db *sql.DB, config client.ConfigProvider) *CustomResourceProvisioner {
 	p := &CustomResourceProvisioner{
-		Provisioners: make(map[string]Provisioner),
+		Provisioners: make(map[string]customresources.Provisioner),
 		client:       http.DefaultClient,
 		sqs:          sqs.New(config),
 	}
@@ -277,9 +85,9 @@ func NewCustomResourceProvisioner(db *sql.DB, config client.ConfigProvider) *Cus
 }
 
 // add adds a custom resource provisioner.
-func (c *CustomResourceProvisioner) add(resourceName string, p Provisioner) {
+func (c *CustomResourceProvisioner) add(resourceName string, p customresources.Provisioner) {
 	// Wrap the provisioner with timeouts.
-	p = withTimeout(p, ProvisioningTimeout, ProvisioningGraceTimeout)
+	p = customresources.WithTimeout(p, ProvisioningTimeout, ProvisioningGraceTimeout)
 	c.Provisioners[resourceName] = p
 }
 
@@ -328,23 +136,23 @@ func (c *CustomResourceProvisioner) Handle(ctx context.Context, message *sqs.Mes
 		return fmt.Errorf("error unmarshalling sqs message body: %v", err)
 	}
 
-	var req Request
+	var req customresources.Request
 	err = json.Unmarshal([]byte(m.Message), &req)
 	if err != nil {
 		return fmt.Errorf("error unmarshalling to cloudformation request: %v", err)
 	}
 
-	resp := NewResponseFromRequest(req)
+	resp := customresources.NewResponseFromRequest(req)
 	resp.PhysicalResourceId, resp.Data, err = c.provision(ctx, m, req)
 	switch err {
 	case nil:
-		resp.Status = StatusSuccess
+		resp.Status = customresources.StatusSuccess
 		logger.Info(ctx, "cloudformation.provision",
 			"request", req,
 			"response", resp,
 		)
 	default:
-		resp.Status = StatusFailed
+		resp.Status = customresources.StatusFailed
 		resp.Reason = err.Error()
 		logger.Error(ctx, "cloudformation.provision.error",
 			"request", req,
@@ -353,31 +161,10 @@ func (c *CustomResourceProvisioner) Handle(ctx context.Context, message *sqs.Mes
 		)
 	}
 
-	raw, err := json.Marshal(resp)
-	if err != nil {
-		return err
-	}
-
-	r, err := http.NewRequest("PUT", req.ResponseURL, bytes.NewReader(raw))
-	if err != nil {
-		return err
-	}
-
-	httpResp, err := c.client.Do(r)
-	if err != nil {
-		return err
-	}
-	defer httpResp.Body.Close()
-	body, _ := ioutil.ReadAll(httpResp.Body)
-
-	if code := httpResp.StatusCode; code/100 != 2 {
-		return fmt.Errorf("unexpected response from pre-signed url: %v: %v", code, string(body))
-	}
-
-	return nil
+	return customresources.SendResponseWithClient(client, req, resp)
 }
 
-func (c *CustomResourceProvisioner) provision(ctx context.Context, m Message, req Request) (string, interface{}, error) {
+func (c *CustomResourceProvisioner) provision(ctx context.Context, m Message, req customresources.Request) (string, interface{}, error) {
 	p, ok := c.Provisioners[req.ResourceType]
 	if !ok {
 		return "", nil, fmt.Errorf("no provisioner for %v", req.ResourceType)
@@ -393,45 +180,4 @@ func (c *CustomResourceProvisioner) provision(ctx context.Context, m Message, re
 	}
 
 	return p.Provision(ctx, req)
-}
-
-// IntValue defines an int64 type that can parse integers as strings from json.
-// It's common to use `Ref`'s inside templates, which means the value of some
-// properties could be a string or an integer.
-type IntValue int64
-
-func intValue(v int64) *IntValue {
-	i := IntValue(v)
-	return &i
-}
-
-func (i *IntValue) UnmarshalJSON(b []byte) error {
-	var si int64
-	if err := json.Unmarshal(b, &si); err == nil {
-		*i = IntValue(si)
-		return nil
-	}
-
-	v, err := strconv.Atoi(string(b[1 : len(b)-1]))
-	if err != nil {
-		return fmt.Errorf("error parsing int from string: %v", err)
-	}
-
-	*i = IntValue(v)
-	return nil
-}
-
-func (i *IntValue) Value() *int64 {
-	if i == nil {
-		return nil
-	}
-	p := int64(*i)
-	return &p
-}
-
-// hashRequest returns a compact unique identifier for the request.
-func hashRequest(r Request) string {
-	h := fnv.New64()
-	h.Write([]byte(fmt.Sprintf("%s.%s", r.StackId, r.RequestId)))
-	return base62.Encode(h.Sum64())
 }
