+++ server/cloudformation/cloudformation_test.go
@@ -6,12 +6,12 @@ import (
 	"io/ioutil"
 	"net/http"
 	"testing"
-	"time"
 
 	"golang.org/x/net/context"
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/sqs"
+	"github.com/remind101/empire/pkg/cloudformation/customresources"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 )
@@ -22,7 +22,7 @@ func TestCustomResourceProvisioner_Handle(t *testing.T) {
 	p := new(mockProvisioner)
 	h := new(mockHTTPClient)
 	c := &CustomResourceProvisioner{
-		Provisioners: map[string]Provisioner{
+		Provisioners: map[string]customresources.Provisioner{
 			"Custom::InstancePort": p,
 		},
 		client: h,
@@ -43,8 +43,8 @@ func TestCustomResourceProvisioner_Handle(t *testing.T) {
 }`),
 	}
 
-	p.On("Provision", Request{
-		RequestType:       Create,
+	p.On("Provision", customresources.Request{
+		RequestType:       customresources.Create,
 		ResponseURL:       "https://cloudformation-custom-resource-response-useast1.s3.amazonaws.com/arn%3Aaws%3Acloudformation%3Aus-east-1%3A066251891493%3Astack/foo/70213b00-0e74-11e6-b4fb-500c28680ac6%7CwebInstancePort%7Cdaf3f3f9-79a1-4049-823e-09544e582b06?AWSAccessKeyId=AKIAJNXHFR7P7YGKLDPQ&Expires=1461987599&Signature=EqV%2BqIUAsZPz5Q%2F%2B75Guvn%2BNREU%3D",
 		StackId:           "arn:aws:cloudformation:us-east-1:066251891493:stack/foo/70213b00-0e74-11e6-b4fb-500c28680ac6",
 		RequestId:         "daf3f3f9-79a1-4049-823e-09544e582b06",
@@ -55,8 +55,8 @@ func TestCustomResourceProvisioner_Handle(t *testing.T) {
 		},
 	}).Return("9001", map[string]int64{"InstancePort": 9001}, nil)
 
-	raw, err := json.Marshal(Response{
-		Status:             StatusSuccess,
+	raw, err := json.Marshal(customresources.Response{
+		Status:             customresources.StatusSuccess,
 		PhysicalResourceId: "9001",
 		StackId:            "arn:aws:cloudformation:us-east-1:066251891493:stack/foo/70213b00-0e74-11e6-b4fb-500c28680ac6",
 		RequestId:          "daf3f3f9-79a1-4049-823e-09544e582b06",
@@ -74,66 +74,11 @@ func TestCustomResourceProvisioner_Handle(t *testing.T) {
 	assert.NoError(t, err)
 }
 
-func TestIntValue(t *testing.T) {
-	type foo struct {
-		I IntValue `json:"I"`
-	}
-
-	tests := []struct {
-		in  []byte
-		out foo
-	}{
-		{[]byte(`{"I": 1}`), foo{I: 1}},
-		{[]byte(`{"I": "1"}`), foo{I: 1}},
-	}
-
-	for _, tt := range tests {
-		var i foo
-		err := json.Unmarshal(tt.in, &i)
-		assert.NoError(t, err)
-		assert.Equal(t, tt.out, i)
-	}
-}
-
-func TestWithTimeout_NoTimeout(t *testing.T) {
-	m := new(mockProvisioner)
-	p := withTimeout(m, time.Second, time.Second)
-
-	m.On("Provision", Request{}).Return("id", nil, nil)
-
-	p.Provision(ctx, Request{})
-}
-
-func TestWithTimeout_Timeout_Cleanup(t *testing.T) {
-	m := new(mockProvisioner)
-	p := withTimeout(m, time.Millisecond*500, time.Millisecond*500)
-
-	m.On("Provision", Request{}).Return("id", nil, nil).Run(func(mock.Arguments) {
-		time.Sleep(time.Millisecond * 750)
-	})
-
-	id, _, err := p.Provision(ctx, Request{})
-	assert.NoError(t, err)
-	assert.Equal(t, "id", id)
-}
-
-func TestWithTimeout_GraceTimeout(t *testing.T) {
-	m := new(mockProvisioner)
-	p := withTimeout(m, time.Millisecond*500, time.Millisecond*500)
-
-	m.On("Provision", Request{}).Return("id", nil, nil).Run(func(mock.Arguments) {
-		time.Sleep(time.Millisecond * 1500)
-	})
-
-	_, _, err := p.Provision(ctx, Request{})
-	assert.Equal(t, context.DeadlineExceeded, err)
-}
-
 type mockProvisioner struct {
 	mock.Mock
 }
 
-func (m *mockProvisioner) Provision(_ context.Context, req Request) (string, interface{}, error) {
+func (m *mockProvisioner) Provision(_ context.Context, req customresources.Request) (string, interface{}, error) {
 	args := m.Called(req)
 	return args.String(0), args.Get(1), args.Error(2)
 }
