+++ server/cloudformation/ports.go
@@ -4,6 +4,8 @@ import (
 	"fmt"
 	"strconv"
 
+	"github.com/remind101/empire/pkg/cloudformation/customresources"
+
 	"golang.org/x/net/context"
 )
 
@@ -21,9 +23,9 @@ func (p *InstancePortsProvisioner) Properties() interface{} {
 	return nil
 }
 
-func (p *InstancePortsProvisioner) Provision(_ context.Context, req Request) (id string, data interface{}, err error) {
+func (p *InstancePortsProvisioner) Provision(_ context.Context, req customresources.Request) (id string, data interface{}, err error) {
 	switch req.RequestType {
-	case Create:
+	case customresources.Create:
 		var port int64
 		port, err = p.ports.Get()
 		if err != nil {
@@ -31,7 +33,7 @@ func (p *InstancePortsProvisioner) Provision(_ context.Context, req Request) (id
 		}
 		id = fmt.Sprintf("%d", port)
 		data = map[string]int64{"InstancePort": port}
-	case Delete:
+	case customresources.Delete:
 		port, err2 := strconv.Atoi(req.PhysicalResourceId)
 		if err2 != nil {
 			err = fmt.Errorf("physical resource id should have been a port number: %v", err2)
